requirements.txt
# ── Telegram ──────────────────────────────────────────────────────────────────
python-telegram-bot==21.6

# ── Database ──────────────────────────────────────────────────────────────────
aiosqlite==0.20.0

# ── Data processing ───────────────────────────────────────────────────────────
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# ── Text / markup ─────────────────────────────────────────────────────────────
PyYAML==6.0.2
lxml==5.3.0
chardet==5.2.0

# ── Documents (optional — comment out to reduce image size) ───────────────────
pdfplumber==0.11.0
python-docx==1.1.2

# ── Search ────────────────────────────────────────────────────────────────────
rapidfuzz==3.6.1

# ── File watching ─────────────────────────────────────────────────────────────
watchdog==4.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
loguru==0.7.2
rich==13.7.0
humanize==4.9.0
tabulate==0.9.0
cachetools==5.3.2
aiofiles==23.2.1
